refactor: 테스트 fake 객체 패키지를 mock 으로 이동
