+++ src/test/java/roomescape/config/TestPaymentConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import roomescape.domain.payment.PaymentClient;
-import roomescape.support.fake.FakePayment;
+import roomescape.support.mock.FakePayment;
 
 @Configuration
 public class TestPaymentConfig {
